Add: Browse page which gets data from database
Might not be bug free yet

+++ BooksOfBliss/src/main/java/com/group05/booksofbliss/view/CheckoutBackingBean.java
@@ -17,7 +17,7 @@ public class CheckoutBackingBean implements Serializable {
 
     @Inject
     @Param(pathIndex = 0)
-    private Integer listingId;
+    private Long listingId;
 
     @EJB
     private ListingDAO listingDAO;
